chore(geofence): re-run refresh for the new user after a switch-triggered cleanup
The exit cleanup stops all monitored regions when the identity changed
mid-refresh, and the new user's own refresh may already have been dropped
on the gate that refresh held — leaving the device monitoring nothing
until the next launch. After releasing the gate, re-run the refresh for
whoever is signed in now with the operation's coordinates, so a user
switch converges to a registered state. Loop-safe: a retry only re-fires
if the identity changes again during the retry itself.

Moves the refresh tier workers verbatim to GeofenceSyncCoordinator+Refresh
(file was at the 400-line cap); only cleanupIfUserChanged changed, gaining
a Bool return.
